Cleaned up all `emailCustomization` flag conditionals
ref https://linear.app/ghost/issue/PROD-2092/

- the feature is now GA so we can remove all of the flag conditionals to clean up unused code and ensure tests are representative of the GA state
- flag left in place for now to enable testing of any further fixes and polish
  - it's also still required to be passed through to Lexical renderer pending similar flag cleanup there
