+++ apps/admin-x-settings/src/components/settings/email/newsletters/NewsletterPreviewContent.tsx
@@ -1,13 +1,8 @@
-import CoverImage from '../../../../assets/images/user-cover.png';
-import CoverImageLabs from '../../../../assets/images/user-cover-labs.jpg';
-import LatestPosts1 from '../../../../assets/images/latest-posts-1.png';
-import LatestPosts1Labs from '../../../../assets/images/latest-posts-1.jpg';
-import LatestPosts2 from '../../../../assets/images/latest-posts-2.png';
-import LatestPosts2Labs from '../../../../assets/images/latest-posts-2.jpg';
-import LatestPosts3 from '../../../../assets/images/latest-posts-3.png';
-import LatestPosts3Labs from '../../../../assets/images/latest-posts-3.jpg';
+import CoverImage from '../../../../assets/images/user-cover.jpg';
+import LatestPosts1 from '../../../../assets/images/latest-posts-1.jpg';
+import LatestPosts2 from '../../../../assets/images/latest-posts-2.jpg';
+import LatestPosts3 from '../../../../assets/images/latest-posts-3.jpg';
 import clsx from 'clsx';
-import useFeatureFlag from '../../../../hooks/useFeatureFlag';
 import {GhostOrb, Icon} from '@tryghost/admin-x-design-system';
 import {isManagedEmail} from '@tryghost/admin-x-framework/api/config';
 import {useGlobalData} from '../../../providers/GlobalDataProvider';
@@ -103,7 +98,6 @@ const NewsletterPreviewContent: React.FC<{
 }) => {
     const showHeader = headerIcon || headerTitle;
     const {config} = useGlobalData();
-    const hasEmailCustomization = useFeatureFlag('emailCustomization');
 
     const currentDate = new Date().toLocaleDateString('default', {
         year: 'numeric',
@@ -168,51 +162,33 @@ const NewsletterPreviewContent: React.FC<{
                                 <div className="py-3">
                                     {headerIcon && <img alt="" className="mx-auto mb-2 size-10" role="presentation" src={headerIcon} />}
                                     {headerTitle && <h4 className="mb-1 text-center text-[1.6rem] font-bold uppercase leading-tight tracking-tight text-grey-900" style={{color: headerTextColor}}>{headerTitle}</h4>}
-                                    {headerSubtitle && <h5 className="mb-1 text-center text-[1.3rem] font-normal text-grey-700" style={{color: hasEmailCustomization ? secondaryHeaderTextColor : secondaryTextColor}}>{headerSubtitle}</h5>}
+                                    {headerSubtitle && <h5 className="mb-1 text-center text-[1.3rem] font-normal text-grey-700" style={{color: secondaryHeaderTextColor}}>{headerSubtitle}</h5>}
                                 </div>
                             )}
                             {showPostTitleSection && (
                                 <div className={clsx('flex flex-col py-8', titleAlignment === 'center' ? 'items-center' : 'items-start')}>
-                                    {hasEmailCustomization ? (
-                                        <>
-                                            <h2 className={clsx(
-                                                'text-4xl font-bold leading-supertight text-black',
-                                                titleFontCategory === 'serif' && 'font-serif',
-                                                titleFontWeight === 'normal' && 'font-normal',
-                                                titleFontWeight === 'medium' && 'font-medium',
-                                                titleFontWeight === 'semibold' && 'font-semibold',
-                                                titleFontWeight === 'bold' && 'font-bold',
-                                                titleAlignment === 'center' ? 'text-center' : 'text-left',
-                                                showExcerpt ? 'mb-2' : 'mb-8'
-                                            )} style={{color: postTitleColor}}>Your email newsletter</h2>
-                                            {showExcerpt && (
-                                                <p className={excerptClasses} style={{color: headerTextColor}}>A subtitle to highlight key points and engage your readers.</p>
-                                            )}
-                                        </>
-                                    ) : (
-                                        <>
-                                            <h2 className={clsx(
-                                                'text-4xl font-bold leading-supertight text-black',
-                                                titleFontCategory === 'serif' && 'font-serif',
-                                                titleAlignment === 'center' ? 'text-center' : 'text-left',
-                                                showExcerpt ? 'mb-2' : 'mb-8'
-                                            )} style={{color: postTitleColor}}>
-                                                Your email newsletter
-                                            </h2>
-                                            {showExcerpt && (
-                                                <p className={excerptClasses}>A subtitle to highlight key points and engage your readers</p>
-                                            )}
-                                        </>
+                                    <h2 className={clsx(
+                                        'text-4xl font-bold leading-supertight text-black',
+                                        titleFontCategory === 'serif' && 'font-serif',
+                                        titleFontWeight === 'normal' && 'font-normal',
+                                        titleFontWeight === 'medium' && 'font-medium',
+                                        titleFontWeight === 'semibold' && 'font-semibold',
+                                        titleFontWeight === 'bold' && 'font-bold',
+                                        titleAlignment === 'center' ? 'text-center' : 'text-left',
+                                        showExcerpt ? 'mb-2' : 'mb-8'
+                                    )} style={{color: postTitleColor}}>Your email newsletter</h2>
+                                    {showExcerpt && (
+                                        <p className={excerptClasses} style={{color: headerTextColor}}>A subtitle to highlight key points and engage your readers.</p>
                                     )}
                                     <div className={clsx(
                                         'flex w-full justify-between text-center text-md leading-none text-grey-700',
                                         titleAlignment === 'center' ? 'flex-col gap-1' : 'flex-row'
                                     )}>
-                                        <p className="pb-1 text-[1.3rem]" style={{color: hasEmailCustomization ? secondaryHeaderTextColor : secondaryTextColor}}>
+                                        <p className="pb-1 text-[1.3rem]" style={{color: secondaryHeaderTextColor}}>
                                             By {authorPlaceholder}
                                             <span className="before:pl-0.5 before:pr-1 before:content-['•']">{currentDate}</span>
                                         </p>
-                                        <p className="pb-1 text-[1.3rem] underline" style={{color: hasEmailCustomization ? secondaryHeaderTextColor : secondaryTextColor}}><span>View in browser</span></p>
+                                        <p className="pb-1 text-[1.3rem] underline" style={{color: secondaryHeaderTextColor}}><span>View in browser</span></p>
                                     </div>
                                 </div>
                             )}
@@ -221,17 +197,16 @@ const NewsletterPreviewContent: React.FC<{
                             {showFeatureImage && (
                                 <>
                                     <div className={clsx(
-                                        'w-full max-w-[600px] bg-cover bg-no-repeat',
-                                        showPostTitleSection ? '' : 'pt-6',
-                                        hasEmailCustomization ? 'h-[unset]' : 'h-[300px]'
+                                        'h-[unset] w-full max-w-[600px] bg-cover bg-no-repeat',
+                                        showPostTitleSection ? '' : 'pt-6'
                                     )}>
                                         <img alt="Feature" className={clsx(
                                             'min-h-full min-w-full shrink-0',
                                             imageCorners === 'square' && 'rounded-none',
                                             imageCorners === 'rounded' && 'rounded-md'
-                                        )} src={hasEmailCustomization ? CoverImageLabs : CoverImage} />
+                                        )} src={CoverImage} />
                                     </div>
-                                    <div className="mt-1 w-full max-w-[600px] pb-8 text-center text-[1.3rem] text-grey-700" style={{color: hasEmailCustomization ? secondaryHeaderTextColor : secondaryTextColor}}>Feature image caption</div>
+                                    <div className="mt-1 w-full max-w-[600px] pb-8 text-center text-[1.3rem] text-grey-700" style={{color: secondaryHeaderTextColor}}>Feature image caption</div>
                                 </>
                             )}
                         </div>
@@ -244,60 +219,50 @@ const NewsletterPreviewContent: React.FC<{
                                 bodyFontCategory === 'serif' ? 'font-serif text-[1.8rem]' : 'text-[1.7rem] tracking-tight',
                                 (showFeatureImage || showPostTitleSection) ? '' : 'pt-8'
                             )} style={{borderColor: dividerColor}}>
-                                {hasEmailCustomization ? (
-                                    <>
-                                        <p className="mb-6" style={{color: textColor}}>This is what your content will look like when you send one of your posts as an email newsletter to your subscribers.</p>
-                                        <p className="mb-6" style={{color: textColor}}>Over there on the right you&apos;ll see some settings that allow you to customize the look and feel of this template – from colors and typography to layout and buttons – to make it perfectly suited to your brand.</p>
-                                        <p className="mb-[52px]" style={{color: textColor}}>Email templates are exceptionally finnicky to make, but we&apos;ve spent a long time optimising this one to make it work beautifully across devices, email clients and content types. So, you can trust that every email you send with Ghost will look great and work well. Just like the rest of your site.</p>
-                                        <hr className={clsx('my-[52px] border-[#e0e7eb]', dividerStyle === 'dashed' && 'border-dashed', dividerStyle === 'dotted' && 'border-b-2 border-t-0 border-dotted')} style={{borderColor: dividerColor}} />
-                                        <h3
-                                            className={clsx(
-                                                'mb-[13px] text-[2.6rem] leading-supertight',
-                                                titleFontCategory === 'serif' && 'font-serif',
-                                                titleFontCategory === 'sans_serif' && 'font-sans',
-                                                titleFontWeight === 'normal' && 'font-normal',
-                                                titleFontWeight === 'medium' && 'font-medium',
-                                                titleFontWeight === 'semibold' && 'font-semibold',
-                                                titleFontWeight === 'bold' && 'font-bold'
-                                            )}
-                                            style={{color: sectionTitleColor}}>Need inspiration?</h3>
-                                        <p className="mb-[27px]" style={{color: textColor}}>We&apos;ve put together a <a className={clsx(linkStyle === 'underline' && 'underline', linkStyle === 'bold' && 'font-bold')} href="https://ghost.org/help/email-design/" rel="noopener noreferrer" style={{color: linkColor || accentColor}} target="_blank">quick guide</a> that walks through all of the available settings, along with a few examples of what&apos;s possible.</p>
-                                        <a
-                                            className={clsx(
-                                                'inline-block border px-[18px] py-2 font-sans text-[15px]',
-                                                buttonCorners === 'rounded' && 'rounded-[6px]',
-                                                buttonCorners === 'pill' && 'rounded-full',
-                                                buttonCorners === 'square' && 'rounded-none',
-                                                buttonStyle === 'outline'
-                                                    ? 'bg-transparent'
-                                                    : 'border-transparent text-white',
-                                                linkStyle === 'bold' ? 'font-bold' : 'font-semibold'
-                                            )}
-                                            href="https://ghost.org/help/email-design/"
-                                            rel="noopener noreferrer"
-                                            style={
-                                                buttonStyle === 'outline'
-                                                    ? {
-                                                        borderColor: buttonColor || accentColor,
-                                                        color: buttonColor || accentColor
-                                                    }
-                                                    : {
-                                                        backgroundColor: buttonColor || accentColor,
-                                                        color: buttonTextColor
-                                                    }
+                                <p className="mb-6" style={{color: textColor}}>This is what your content will look like when you send one of your posts as an email newsletter to your subscribers.</p>
+                                <p className="mb-6" style={{color: textColor}}>Over there on the right you&apos;ll see some settings that allow you to customize the look and feel of this template – from colors and typography to layout and buttons – to make it perfectly suited to your brand.</p>
+                                <p className="mb-[52px]" style={{color: textColor}}>Email templates are exceptionally finnicky to make, but we&apos;ve spent a long time optimising this one to make it work beautifully across devices, email clients and content types. So, you can trust that every email you send with Ghost will look great and work well. Just like the rest of your site.</p>
+                                <hr className={clsx('my-[52px] border-[#e0e7eb]', dividerStyle === 'dashed' && 'border-dashed', dividerStyle === 'dotted' && 'border-b-2 border-t-0 border-dotted')} style={{borderColor: dividerColor}} />
+                                <h3
+                                    className={clsx(
+                                        'mb-[13px] text-[2.6rem] leading-supertight',
+                                        titleFontCategory === 'serif' && 'font-serif',
+                                        titleFontCategory === 'sans_serif' && 'font-sans',
+                                        titleFontWeight === 'normal' && 'font-normal',
+                                        titleFontWeight === 'medium' && 'font-medium',
+                                        titleFontWeight === 'semibold' && 'font-semibold',
+                                        titleFontWeight === 'bold' && 'font-bold'
+                                    )}
+                                    style={{color: sectionTitleColor}}>Need inspiration?</h3>
+                                <p className="mb-[27px]" style={{color: textColor}}>We&apos;ve put together a <a className={clsx(linkStyle === 'underline' && 'underline', linkStyle === 'bold' && 'font-bold')} href="https://ghost.org/help/email-design/" rel="noopener noreferrer" style={{color: linkColor || accentColor}} target="_blank">quick guide</a> that walks through all of the available settings, along with a few examples of what&apos;s possible.</p>
+                                <a
+                                    className={clsx(
+                                        'inline-block border px-[18px] py-2 font-sans text-[15px]',
+                                        buttonCorners === 'rounded' && 'rounded-[6px]',
+                                        buttonCorners === 'pill' && 'rounded-full',
+                                        buttonCorners === 'square' && 'rounded-none',
+                                        buttonStyle === 'outline'
+                                            ? 'bg-transparent'
+                                            : 'border-transparent text-white',
+                                        linkStyle === 'bold' ? 'font-bold' : 'font-semibold'
+                                    )}
+                                    href="https://ghost.org/help/email-design/"
+                                    rel="noopener noreferrer"
+                                    style={
+                                        buttonStyle === 'outline'
+                                            ? {
+                                                borderColor: buttonColor || accentColor,
+                                                color: buttonColor || accentColor
                                             }
-                                            target="_blank"
-                                        >
-                                            Learn more
-                                        </a>
-                                    </>
-                                ) : (
-                                    <>
-                                        <p className="mb-6" style={{color: textColor}}>This is what your content will look like when you send one of your posts as an email newsletter to your subscribers.</p>
-                                        <p className="mb-6" style={{color: textColor}}>Over there on the right you&apos;ll see some settings that allow you to customize the look and feel of this template to make it perfectly suited to your brand. Email templates are exceptionally finnicky to make, but we&apos;ve spent a long time optimising this one to make it work beautifully across devices, email clients and content types.</p>
-                                        <p style={{color: textColor}}>So, you can trust that every email you send with Ghost will look great and work well. Just like the rest of your site.</p>
-                                    </>
-                                )}
+                                            : {
+                                                backgroundColor: buttonColor || accentColor,
+                                                color: buttonTextColor
+                                            }
+                                    }
+                                    target="_blank"
+                                >
+                                    Learn more
+                                </a>
                             </div>
 
                             {/* Feedback */}
@@ -341,7 +306,7 @@ const NewsletterPreviewContent: React.FC<{
                                             <h4
                                                 className={clsx(
                                                     'mt-0.5 text-[1.9rem] text-black',
-                                                    hasEmailCustomization && titleFontCategory === 'serif' && 'font-serif',
+                                                    titleFontCategory === 'serif' && 'font-serif',
                                                     titleFontWeight === 'normal' && 'font-normal',
                                                     titleFontWeight === 'medium' && 'font-medium',
                                                     titleFontWeight === 'semibold' && 'font-semibold',
@@ -354,7 +319,7 @@ const NewsletterPreviewContent: React.FC<{
                                             <img alt="Latest post" className={clsx(
                                                 imageCorners === 'square' && 'rounded-none',
                                                 imageCorners === 'rounded' && 'rounded-md'
-                                            )} src={hasEmailCustomization ? LatestPosts1Labs : LatestPosts1} />
+                                            )} src={LatestPosts1} />
                                         </div>
                                     </div>
                                     <div className="flex justify-between gap-4 py-2">
@@ -362,7 +327,7 @@ const NewsletterPreviewContent: React.FC<{
                                             <h4
                                                 className={clsx(
                                                     'mt-0.5 text-[1.9rem] text-black',
-                                                    hasEmailCustomization && titleFontCategory === 'serif' && 'font-serif',
+                                                    titleFontCategory === 'serif' && 'font-serif',
                                                     titleFontWeight === 'normal' && 'font-normal',
                                                     titleFontWeight === 'medium' && 'font-medium',
                                                     titleFontWeight === 'semibold' && 'font-semibold',
@@ -374,7 +339,7 @@ const NewsletterPreviewContent: React.FC<{
                                             <img alt="Latest post" className={clsx(
                                                 imageCorners === 'square' && 'rounded-none',
                                                 imageCorners === 'rounded' && 'rounded-md'
-                                            )} src={hasEmailCustomization ? LatestPosts2Labs : LatestPosts2} />
+                                            )} src={LatestPosts2} />
                                         </div>
                                     </div>
                                     <div className="flex justify-between gap-4 py-2">
@@ -382,7 +347,7 @@ const NewsletterPreviewContent: React.FC<{
                                             <h4
                                                 className={clsx(
                                                     'mt-0.5 text-[1.9rem] text-black',
-                                                    hasEmailCustomization && titleFontCategory === 'serif' && 'font-serif',
+                                                    titleFontCategory === 'serif' && 'font-serif',
                                                     titleFontWeight === 'normal' && 'font-normal',
                                                     titleFontWeight === 'medium' && 'font-medium',
                                                     titleFontWeight === 'semibold' && 'font-semibold',
@@ -394,7 +359,7 @@ const NewsletterPreviewContent: React.FC<{
                                             <img alt="Latest post" className={clsx(
                                                 imageCorners === 'square' && 'rounded-none',
                                                 imageCorners === 'rounded' && 'rounded-md'
-                                            )} src={hasEmailCustomization ? LatestPosts3Labs : LatestPosts3} />
+                                            )} src={LatestPosts3} />
                                         </div>
                                     </div>
                                 </div>
