+++ ghost/core/core/server/services/email-service/EmailRenderer.js
@@ -325,34 +325,31 @@ class EmailRenderer {
             postUrl
         };
 
-        const labs = this.getLabs();
         const accentColor = this.#getAccentColor();
         const accentContrastColor = this.#getAccentContrastColor();
 
-        if (labs?.isSet('emailCustomization')) {
-            renderOptions.design = {
-                accentColor,
-                accentContrastColor,
-                backgroundColor: newsletter?.get('background_color'),
-                backgroundIsDark: this.#checkIfBackgroundIsDark(newsletter),
-                headerBackgroundColor: this.#getHeaderBackgroundColor(newsletter, accentColor),
-                buttonCorners: newsletter?.get('button_corners'),
-                buttonStyle: newsletter?.get('button_style'),
-                titleFontWeight: newsletter?.get('title_font_weight'),
-                linkStyle: newsletter?.get('link_style'),
-                imageCorners: newsletter?.get('image_corners'),
-                postTitleColor: newsletter?.get('post_title_color'),
-                sectionTitleColor: newsletter?.get('section_title_color'),
-                linkColor: newsletter?.get('link_color'),
-                // TODO:
-                // if the other options above have default or calculated values we
-                // should follow the same pattern as the options below to avoid
-                //duplicating magic values or logic in renderers
-                dividerColor: this.#getDividerColor(newsletter),
-                buttonColor: this.#getButtonColor(newsletter, accentColor),
-                buttonTextColor: this.#getButtonTextColor(newsletter, accentColor)
-            };
-        }
+        renderOptions.design = {
+            accentColor,
+            accentContrastColor,
+            backgroundColor: newsletter?.get('background_color'),
+            backgroundIsDark: this.#checkIfBackgroundIsDark(newsletter),
+            headerBackgroundColor: this.#getHeaderBackgroundColor(newsletter, accentColor),
+            buttonCorners: newsletter?.get('button_corners'),
+            buttonStyle: newsletter?.get('button_style'),
+            titleFontWeight: newsletter?.get('title_font_weight'),
+            linkStyle: newsletter?.get('link_style'),
+            imageCorners: newsletter?.get('image_corners'),
+            postTitleColor: this.#getPostTitleColor(newsletter, accentColor),
+            sectionTitleColor: newsletter?.get('section_title_color'),
+            linkColor: newsletter?.get('link_color'),
+            // TODO:
+            // if the other options above have default or calculated values we
+            // should follow the same pattern as the options below to avoid
+            //duplicating magic values or logic in renderers
+            dividerColor: this.#getDividerColor(newsletter),
+            buttonColor: this.#getButtonColor(newsletter, accentColor),
+            buttonTextColor: this.#getButtonTextColor(newsletter, accentColor)
+        };
 
         let html;
         if (post.get('lexical')) {
@@ -862,9 +859,6 @@ class EmailRenderer {
 
         // Actual template
         let templateName = 'template.hbs';
-        if (labs?.isSet('emailCustomization')) {
-            templateName = 'template-emailCustomization.hbs';
-        }
         const htmlTemplateSource = await fs.readFile(path.join(__dirname, './email-templates/', templateName), 'utf8');
         this.#renderTemplate = this.#handlebars.compile(Buffer.from(htmlTemplateSource).toString());
 
@@ -975,11 +969,6 @@ class EmailRenderer {
     }
 
     #getSectionTitleColor(newsletter, accentColor) {
-        const labs = this.getLabs();
-        if (!labs?.isSet('emailCustomization')) {
-            return null;
-        }
-
         /** @type {'accent' | string | null} */
         const value = newsletter.get('section_title_color');
 
@@ -1002,11 +991,6 @@ class EmailRenderer {
             bold: '700'
         };
 
-        const labs = this.getLabs();
-        if (!labs?.isSet('emailCustomization')) {
-            return weights.bold;
-        }
-
         /** @type {'normal' | 'medium' | 'semibold' | 'bold' | string | null} */
         const settingValue = newsletter.get('title_font_weight');
 
@@ -1016,7 +1000,7 @@ class EmailRenderer {
     #getTitleStrongWeight(titleWeight) {
         const numericWeight = parseInt(titleWeight, 10);
 
-        if (isNaN(numericWeight) || !this.#labs?.isSet('emailCustomization')) {
+        if (isNaN(numericWeight)) {
             return '800';
         }
 
@@ -1038,20 +1022,16 @@ class EmailRenderer {
     }
 
     #getDividerColor(newsletter) {
-        const labs = this.getLabs();
+        const value = newsletter?.get('divider_color');
 
-        if (labs?.isSet('emailCustomization')) {
-            const value = newsletter?.get('divider_color');
-
-            if (value === 'accent') {
-                return this.#getAccentColor();
-            } else if (VALID_HEX_REGEX.test(value)) {
-                return value;
-            }
+        if (value === 'accent') {
+            return this.#getAccentColor();
+        } else if (VALID_HEX_REGEX.test(value)) {
+            return value;
+        } else {
+            // value === 'light'/missing/invalid
+            return '#e0e7eb';
         }
-
-        // value === 'light'/missing/invalid
-        return '#e0e7eb';
     }
 
     #getLinkColor(newsletter, accentColor) {
@@ -1126,49 +1106,33 @@ class EmailRenderer {
      * @private
      */
     async getTemplateData({post, newsletter, html, addPaywall, segment}) {
-        const labs = this.getLabs();
-
         const accentColor = this.#getAccentColor();
         const accentContrastColor = this.#getAccentContrastColor();
 
-        // TODO: remove adjusted accent colors when emailCustomization flag is cleaned up
-        let adjustedAccentColor;
-        let adjustedAccentContrastColor;
-        try {
-            adjustedAccentColor = accentColor && darkenToContrastThreshold(accentColor, '#ffffff', 2).hex();
-            adjustedAccentContrastColor = accentColor && textColorForBackgroundColor(adjustedAccentColor).hex();
-        } catch (e) {
-            logging.error(e);
-            adjustedAccentColor = accentColor;
-        }
-
-        const hasEmailCustomization = labs.isSet('emailCustomization');
-
+        // TODO: remove passthrough of accent color to getters
         const backgroundColor = this.#getBackgroundColor(newsletter);
         const backgroundIsDark = this.#checkIfBackgroundIsDark(newsletter);
         const postTitleColor = this.#getPostTitleColor(newsletter, accentColor);
         const titleWeight = this.#getTitleWeight(newsletter);
         const titleStrongWeight = this.#getTitleStrongWeight(titleWeight);
         const textColor = textColorForBackgroundColor(backgroundColor).hex(); // this is used by the header background color so keeping it separate from the content text color
-        const linkColor = hasEmailCustomization ? this.#getLinkColor(newsletter, accentColor) : backgroundIsDark ? '#ffffff' : accentColor;
-        const hasRoundedImageCorners = hasEmailCustomization ? this.#getImageCorners(newsletter) : false;
-        const sectionTitleColor = hasEmailCustomization ? this.#getSectionTitleColor(newsletter, accentColor) : null;
+        const linkColor = this.#getLinkColor(newsletter, accentColor);
+        const hasRoundedImageCorners = this.#getImageCorners(newsletter);
+        const sectionTitleColor = this.#getSectionTitleColor(newsletter, accentColor);
         const dividerColor = this.#getDividerColor(newsletter);
-        const buttonColor = this.#getButtonColor(newsletter, adjustedAccentColor);
-        const buttonTextColor = this.#getButtonTextColor(newsletter, adjustedAccentColor);
+        const buttonColor = this.#getButtonColor(newsletter, accentColor);
+        const buttonTextColor = this.#getButtonTextColor(newsletter, accentColor);
         const headerBackgroundColor = this.#getHeaderBackgroundColor(newsletter, accentColor);
         const headerBackgroundIsDark = textColorForBackgroundColor(headerBackgroundColor || backgroundColor).hex().toLowerCase() === '#ffffff';
 
         let buttonBorderRadius = '6px';
-        if (hasEmailCustomization) {
-            if (newsletter.get('button_corners') === 'square') {
-                buttonBorderRadius = '0';
-            } else if (newsletter.get('button_corners') === 'pill') {
-                buttonBorderRadius = '9999px';
-            }
+        if (newsletter.get('button_corners') === 'square') {
+            buttonBorderRadius = '0';
+        } else if (newsletter.get('button_corners') === 'pill') {
+            buttonBorderRadius = '9999px';
         }
 
-        const hasOutlineButtons = hasEmailCustomization && newsletter.get('button_style') === 'outline';
+        const hasOutlineButtons = newsletter.get('button_style') === 'outline';
 
         const {href: headerImage, width: headerImageWidth} = await this.limitImageWidth(newsletter.get('header_image'));
         const {href: postFeatureImage, width: postFeatureImageWidth, height: postFeatureImageHeight} = await this.limitImageWidth(post.get('feature_image'));
@@ -1258,7 +1222,7 @@ class EmailRenderer {
         const titleAlignment = newsletter.get('title_alignment');
         const showFeatureImage = newsletter.get('show_feature_image') && !!postFeatureImage;
 
-        const linkStyle = (hasEmailCustomization && newsletter.get('link_style')) || 'underline';
+        const linkStyle = newsletter.get('link_style') || 'underline';
 
         const data = {
             site: {
@@ -1307,10 +1271,8 @@ class EmailRenderer {
             latestPostsHasImages,
 
             //CSS
-            accentColor: accentColor, // default to #15212A
+            accentColor, // default to #15212A
             accentContrastColor,
-            adjustedAccentColor: adjustedAccentColor || '#3498db', // default to #3498db
-            adjustedAccentContrastColor: adjustedAccentContrastColor || '#ffffff', // default to #ffffff
             showBadge: newsletter.get('show_badge'),
             backgroundColor,
             backgroundIsDark,
@@ -1347,8 +1309,7 @@ class EmailRenderer {
                     'post-title-with-excerpt': post.get('custom_excerpt'),
                     'post-title-no-excerpt': !post.get('custom_excerpt'),
                     'post-title-serif': titleFont === 'serif',
-                    'post-title-left': titleAlignment === 'left',
-                    'post-title-color': hasEmailCustomization
+                    'post-title-left': titleAlignment === 'left'
                 }),
                 titleLink: clsx('post-title-link', {
                     'post-title-link-left': titleAlignment === 'left'
