+++ ghost/core/core/server/services/email-service/email-templates/partials/paywall.hbs
@@ -1,63 +1,36 @@
-{{#hasFeature "emailCustomization"}}
-    <div class="kg-paywall align-center">
-        <table class="kg-paywall-hr" role="presentation" width="100%" border="0" cellpadding="0" cellspacing="0">
-            <tbody>
-                <tr>
-                    <td>
-                        <!--[if !mso]><!-- -->
-                        <hr style="display: none;" />
-                        <!--<![endif]-->
-                        <table class="kg-hr" role="presentation" border="0" cellpadding="0" cellspacing="0">
-                            <tbody>
-                                <tr>
-                                    <td>&nbsp;</td>
-                                </tr>
-                            </tbody>
-                        </table>
-                    </td>
-                </tr>
-            </tbody>
-        </table>
-        <h3>{{t 'Upgrade to continue reading.'}}</h3>
-        <p>{{{t 'Become a paid member of {site} to get access to all premium content.' site=site.title}}}</p>
-            <table border="0" cellpadding="0" cellspacing="0">
-                <tr>
-                    <td>
-                        {{#hasFeature "emailCustomization"}}
-                        <table class="btn" border="0" cellspacing="0" cellpadding="0" align="center">
-                        {{else}}
-                        <table class="btn btn-accent" border="0" cellspacing="0" cellpadding="0" align="center">
-                        {{/hasFeature}}
+<div class="kg-paywall align-center">
+    <table class="kg-paywall-hr" role="presentation" width="100%" border="0" cellpadding="0" cellspacing="0">
+        <tbody>
+            <tr>
+                <td>
+                    <!--[if !mso]><!-- -->
+                    <hr style="display: none;" />
+                    <!--<![endif]-->
+                    <table class="kg-hr" role="presentation" border="0" cellpadding="0" cellspacing="0">
+                        <tbody>
                             <tr>
-                                <td align="center">
-                                    <a href="{{paywall.signupUrl}}">{{t 'Upgrade'}}</a>
-                                </td>
+                                <td>&nbsp;</td>
                             </tr>
-                        </table>
-                    </td>
-                </tr>
-            </table>
-        <p></p>
-    </div>
-{{else}}
-    <div class="align-center" style="text-align: center;">
-        <hr
-            style="position: relative; display: block; width: 100%; margin: 3em 0; padding: 0; height: 1px; border: 0; border-top: 1px solid #e0e7eb;">
-        <h3
-            style="margin-top: 0; font-family: -apple-system, BlinkMacSystemFont, Roboto, Helvetica, Arial, sans-serif; line-height: 1.11em; font-weight: 700; text-rendering: optimizeLegibility; margin: 1.5em 0 0.5em 0; text-wrap: pretty;">
-            {{t 'Upgrade to continue reading.'}}</h3>
-        <p style="margin: 0 auto 1.5em auto; line-height: 1.6em; max-width: 440px; text-wrap: pretty;">{{{t 'Become a paid member of {site} to get access to all premium content.' site=site.title}}}</p>
-            <div class="btn btn-accent" style="box-sizing: border-box; width: 100%; display: table;">
-                <table border="0" cellspacing="0" cellpadding="0" align="center" style="border-collapse: separate; mso-table-lspace: 0pt; mso-table-rspace: 0pt; width: auto;">
-                    <tbody>
+                        </tbody>
+                    </table>
+                </td>
+            </tr>
+        </tbody>
+    </table>
+    <h3>{{t 'Upgrade to continue reading.'}}</h3>
+    <p>{{{t 'Become a paid member of {site} to get access to all premium content.' site=site.title}}}</p>
+        <table border="0" cellpadding="0" cellspacing="0">
+            <tr>
+                <td>
+                    <table class="btn" border="0" cellspacing="0" cellpadding="0" align="center">
                         <tr>
-                            <td align="center" valign="top" bgcolor="{{accentColor}}" style="vertical-align: top; text-align: center; border-radius: 5px;">
-                                <a href="{{paywall.signupUrl}}" style="overflow-wrap: anywhere; border: solid 1px #3498db; border-radius: 5px; box-sizing: border-box; cursor: pointer; display: inline-block; font-size: 14px; font-weight: bold; margin: 0; padding: 12px 25px; text-decoration: none; background-color: {{accentColor}}; border-color: {{accentColor}}; color: #FFFFFF;" target="_blank">{{t 'Upgrade'}}</a>
+                            <td align="center">
+                                <a href="{{paywall.signupUrl}}">{{t 'Upgrade'}}</a>
                             </td>
                         </tr>
-                    </tbody>
-                </table>
-            </div>
-        <p style="margin: 0 0 1.5em 0; line-height: 1.6em;"></p>
-    </div>
-{{/hasFeature}}
+                    </table>
+                </td>
+            </tr>
+        </table>
+    <p></p>
+</div>
