+++ ghost/core/core/server/services/koenig/node-renderers/button-renderer.js
@@ -36,15 +36,13 @@ function frontendTemplate(node, document) {
 function emailTemplate(node, options, document) {
     const {buttonUrl, buttonText} = node;
 
-    let cardHtml;
-    if (options.feature?.emailCustomization) {
-        const buttonHtml = renderEmailButton({
-            alignment: node.alignment,
-            url: buttonUrl,
-            text: buttonText
-        });
+    const buttonHtml = renderEmailButton({
+        alignment: node.alignment,
+        url: buttonUrl,
+        text: buttonText
+    });
 
-        cardHtml = html`
+    const cardHtml = html`
         <table class="kg-card kg-button-card" border="0" cellpadding="0" cellspacing="0">
             <tbody>
                 <tr>
@@ -54,28 +52,11 @@ function emailTemplate(node, options, document) {
                 </tr>
             </tbody>
         </table>
-        `;
+    `;
 
-        const element = document.createElement('div');
-        element.innerHTML = cardHtml;
-        return {element, type: 'inner'};
-    } else {
-        cardHtml = html`
-        <div class="btn btn-accent">
-            <table border="0" cellspacing="0" cellpadding="0" align="${node.alignment}">
-                <tr>
-                    <td align="center">
-                        <a href="${buttonUrl}">${buttonText}</a>
-                    </td>
-                </tr>
-            </table>
-        </div>
-        `;
-
-        const element = document.createElement('p');
-        element.innerHTML = cardHtml;
-        return {element};
-    }
+    const element = document.createElement('div');
+    element.innerHTML = cardHtml;
+    return {element, type: 'inner'};
 }
 
 function getCardClasses(node) {
