+++ ghost/core/test/unit/server/services/koenig/node-renderers/button-renderer.test.js
@@ -49,52 +49,7 @@ describe('services/koenig/node-renderers/button-renderer', function () {
 
     describe('email', function () {
         it('matches snapshot for default test data', function () {
-            const result = renderForEmail(getTestData());
-
-            assert.ok(result.html);
-
-            assertPrettifiesTo(result.html, html`
-                <p></p>
-                <div class="btn btn-accent">
-                    <table border="0" cellspacing="0" cellpadding="0" align="center">
-                        <tbody>
-                            <tr>
-                                <td align="center"><a href="http://blog.com/post1">click me</a></td>
-                            </tr>
-                        </tbody>
-                    </table>
-                </div>
-                <p></p>
-            `);
-        });
-
-        // TODO: This is a bug in the email renderer, it should render the default button text
-        it('renders empty button with missing buttonText', function () {
-            const result = renderForEmail(getTestData({buttonText: ''}));
-            assertPrettifiesTo(result.html, html`
-                <p></p>
-                <div class="btn btn-accent">
-                    <table border="0" cellspacing="0" cellpadding="0" align="center">
-                        <tbody>
-                            <tr>
-                                <td align="center"><a href="http://blog.com/post1"></a></td>
-                            </tr>
-                        </tbody>
-                    </table>
-                </div>
-                <p></p>
-            `);
-        });
-
-        it('renders nothing with a missing buttonUrl', function () {
-            const result = renderForEmail(getTestData({buttonUrl: ''}));
-            assert.equal(result.html, '');
-        });
-    });
-
-    describe('email (emailCustomization)', function () {
-        it('matches snapshot for default test data', function () {
-            const result = renderForEmail(getTestData(), {feature: {emailCustomization: true}});
+            const result = renderForEmail(getTestData(), {feature: {}});
 
             assertPrettifiesTo(result.html, html`
                 <table class="kg-card kg-button-card" border="0" cellpadding="0" cellspacing="0">
@@ -115,6 +70,35 @@ describe('services/koenig/node-renderers/button-renderer', function () {
                     </tbody>
                 </table>
             `);
+        });
+
+        // TODO: This is a bug in the email renderer, it should render the default button text
+        it('renders empty button with missing buttonText', function () {
+            const result = renderForEmail(getTestData({buttonText: ''}));
+            assertPrettifiesTo(result.html, html`
+                <table class="kg-card kg-button-card" border="0" cellpadding="0" cellspacing="0">
+                    <tbody>
+                        <tr>
+                            <td class="kg-card-spacing">
+                                <table class="btn" border="0" cellspacing="0" cellpadding="0" align="center">
+                                    <tbody>
+                                        <tr>
+                                            <td align="center">
+                                                <a href="http://blog.com/post1"></a>
+                                            </td>
+                                        </tr>
+                                    </tbody>
+                                </table>
+                            </td>
+                        </tr>
+                    </tbody>
+                </table>
+            `);
+        });
+
+        it('renders nothing with a missing buttonUrl', function () {
+            const result = renderForEmail(getTestData({buttonUrl: ''}));
+            assert.equal(result.html, '');
         });
     });
 });
