+++ ghost/core/test/unit/server/services/koenig/node-renderers/call-to-action-renderer.test.js
@@ -73,74 +73,7 @@ describe('services/koenig/node-renderers/call-to-action-renderer', function () {
 
     describe('email', function () {
         it('matches snapshot for default test data', function () {
-            const result = renderForEmail(getTestData());
-
-            assert.ok(result.html);
-
-            assertPrettifiesTo(result.html, html`
-                <table class="kg-card kg-cta-card kg-cta-bg-none kg-cta-minimal kg-cta-has-img" border="0" cellpadding="0" cellspacing="0" width="100%">
-                    <tbody>
-                        <tr>
-                            <td>
-                                <table border="0" cellpadding="0" cellspacing="0" width="100%">
-                                    <tbody>
-                                        <tr>
-                                            <td class="kg-cta-sponsor-label">
-                                                <p><span style="white-space: pre-wrap">SPONSORED</span></p>
-                                            </td>
-                                        </tr>
-                                    </tbody>
-                                </table>
-                            </td>
-                        </tr>
-                        <tr>
-                            <td class="kg-cta-content">
-                                <table border="0" cellpadding="0" cellspacing="0" width="100%" class="kg-cta-content-wrapper">
-                                    <tbody>
-                                        <tr>
-                                            <td class="kg-cta-image-container" width="64">
-                                                <a href="http://blog.com/post1">
-                                                    <img src="http://blog.com/image1.jpg" alt="CTA Image" class="kg-cta-image" width="64" height="64"/>
-                                                </a>
-                                            </td>
-                                            <td class="kg-cta-content-inner">
-                                                <table border="0" cellpadding="0" cellspacing="0" width="100%">
-                                                    <tbody>
-                                                        <tr>
-                                                            <td class="kg-cta-text">This is a cool advertisement</td>
-                                                        </tr>
-                                                        <tr>
-                                                            <td class="kg-cta-button-container">
-                                                                <table border="0" cellpadding="0" cellspacing="0" class="kg-cta-button-wrapper">
-                                                                    <tbody>
-                                                                        <tr>
-                                                                            <td class="" style="background-color: #000000; color: #ffffff;">
-                                                                                <a href="http://blog.com/post1" class="kg-cta-button" style="background-color: #000000; color: #ffffff;">
-                                                                                    click me
-                                                                                </a>
-                                                                            </td>
-                                                                        </tr>
-                                                                    </tbody>
-                                                                </table>
-                                                            </td>
-                                                        </tr>
-                                                    </tbody>
-                                                </table>
-                                            </td>
-                                        </tr>
-                                    </tbody>
-                                </table>
-                            </td>
-                        </tr>
-                    </tbody>
-                </table>
-            `);
-        });
-    });
-
-    describe('email (emailCustomization)', function () {
-        it('matches snapshot for default test data', function () {
-            const result = renderForEmail(getTestData(), {feature: {emailCustomization: true}});
+            const result = renderForEmail(getTestData(), {feature: {}});
 
             assert.ok(result.html);
 
@@ -206,7 +139,7 @@ describe('services/koenig/node-renderers/call-to-action-renderer', function () {
 
         it('handles accent button color', function () {
             const data = getTestData({buttonColor: 'accent'});
-            const result = renderForEmail(data, {feature: {emailCustomization: true}});
+            const result = renderForEmail(data, {feature: {}});
 
             assertPrettifiedIncludes(result.html, html`
                 <table class="btn btn-accent" border="0" cellspacing="0" cellpadding="0">
@@ -225,7 +158,7 @@ describe('services/koenig/node-renderers/call-to-action-renderer', function () {
 
         it('handles outline buttons', function () {
             const data = getTestData();
-            const result = renderForEmail(data, {design: {buttonStyle: 'outline'}, feature: {emailCustomization: true}});
+            const result = renderForEmail(data, {design: {buttonStyle: 'outline'}, feature: {}});
 
             assertPrettifiedIncludes(result.html, html`
                 <table class="btn" border="0" cellspacing="0" cellpadding="0">
@@ -249,7 +182,7 @@ describe('services/koenig/node-renderers/call-to-action-renderer', function () {
             });
             const result = renderForEmail(data, {
                 design: {backgroundIsDark: true},
-                feature: {emailCustomization: true}
+                feature: {}
             });
 
             assertPrettifiedIncludes(result.html, html`
@@ -274,7 +207,7 @@ describe('services/koenig/node-renderers/call-to-action-renderer', function () {
             });
             const result = renderForEmail(data, {
                 design: {backgroundIsDark: true},
-                feature: {emailCustomization: true}
+                feature: {}
             });
 
             assertPrettifiedIncludes(result.html, html`
@@ -299,7 +232,7 @@ describe('services/koenig/node-renderers/call-to-action-renderer', function () {
             });
             const result = renderForEmail(data, {
                 design: {backgroundIsDark: true},
-                feature: {emailCustomization: true}
+                feature: {}
             });
 
             assertPrettifiedIncludes(result.html, html`
@@ -324,7 +257,7 @@ describe('services/koenig/node-renderers/call-to-action-renderer', function () {
             });
             const result = renderForEmail(data, {
                 design: {backgroundIsDark: true},
-                feature: {emailCustomization: true}
+                feature: {}
             });
 
             assertPrettifiedIncludes(result.html, html`
@@ -349,7 +282,7 @@ describe('services/koenig/node-renderers/call-to-action-renderer', function () {
             });
             const result = renderForEmail(data, {
                 design: {backgroundIsDark: true},
-                feature: {emailCustomization: true}
+                feature: {}
             });
 
             assertPrettifiedIncludes(result.html, html`
@@ -374,7 +307,7 @@ describe('services/koenig/node-renderers/call-to-action-renderer', function () {
             });
             const result = renderForEmail(data, {
                 design: {backgroundIsDark: false},
-                feature: {emailCustomization: true}
+                feature: {}
             });
 
             assertPrettifiedIncludes(result.html, html`
@@ -399,7 +332,7 @@ describe('services/koenig/node-renderers/call-to-action-renderer', function () {
             });
             const result = renderForEmail(data, {
                 design: {backgroundIsDark: true},
-                feature: {emailCustomization: true}
+                feature: {}
             });
 
             assertPrettifiedIncludes(result.html, html`
